requirements.txt
# requirements.txt for DefectDojo using Python 3.x
bleach==6.3.0
bleach[css]
celery==5.6.1
defusedxml==0.7.1
django_celery_results==2.6.0
django-auditlog==3.2.1
django-pghistory==3.9.1
django-dbbackup==5.1.0
django-environ==0.12.0
django-filter==25.2
django-imagekit==6.0.0
django-multiselectfield==1.0.1
django-polymorphic==4.5.2
django-crispy-forms==2.5
django_extensions==4.1
django-slack==5.19.0
django-watson==1.6.3
django-prometheus==2.4.1
Django==5.2.9
django-single-session==0.2.0
djangorestframework==3.16.1
html2text==2025.4.15
humanize==4.15.0
jira==3.10.5
PyGithub==2.8.1
lxml==6.0.2
Markdown==3.10
openpyxl==3.1.5
Pillow==12.1.0  # required by django-imagekit
psycopg[c]==3.3.2
cryptography==46.0.3
python-dateutil==2.9.0.post0
redis==7.1.0
requests==2.32.5
sqlalchemy==2.0.45  # Required by Celery broker transport
urllib3==2.6.3
uWSGI==2.0.31
vobject==0.9.9
whitenoise==5.2.0
titlecase==2.4.1
social-auth-app-django==5.6.0
social-auth-core==4.8.3
gitpython==3.1.46
python-gitlab==7.1.0
cpe==1.3.1
packageurl-python==0.17.6
django-crum==0.7.9
JSON-log-formatter==1.1.1
django-split-settings==1.3.2
# do not upgrade to 2.1.1 - https://github.com/DefectDojo/django-DefectDojo/issues/12918
# use fork with django 5.2 fixes, but based on 2.1.0
git+https://github.com/valentijnscholten/django-tagulous.git@2b514f9140acfce608238d1426d864185b3c60a2#egg=django-tagulous
PyJWT==2.10.1
cvss==3.6
django-fieldsignals==0.8.0
hyperlink==21.0.0
djangosaml2==1.11.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
django-ratelimit==4.1.0
argon2-cffi==25.1.0
blackduck==1.1.3
pycurl==7.45.7  # Required for Celery Broker AWS (SQS) support
boto3==1.41.5  # Required for Celery Broker AWS (SQS) support
netaddr==1.3.0
vulners==3.1.3
fontawesomefree==6.6.0
PyYAML==6.0.3
pyopenssl==25.3.0
parameterized==0.9.0
setuptools==80.9.0
django-github-app==0.9.0
